requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
requests==2.32.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
google-generativeai==0.7.2
httpx==0.27.0
python-multipart==0.0.9
